Extract Text/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.46.0
pillow>=9.0.0
accelerate>=0.20.0
qwen-vl-utils
